Added headers filtering to flooding protection
