+++ kura/org.eclipse.kura.net.admin.firewall/src/main/java/org/eclipse/kura/net/admin/AbstractFirewallConfigurationServiceImpl.java
@@ -14,6 +14,7 @@ package org.eclipse.kura.net.admin;
 
 import java.net.UnknownHostException;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -354,4 +355,17 @@ public abstract class AbstractFirewallConfigurationServiceImpl<U extends IPAddre
         return this.firewall.getPortForwardRules();
     }
 
+    public void addFloodingProtectionRules(Set<String> floodingRules) {
+        // The flooding protection rules are applied only to the mangle table.
+        addFloodingProtectionRules(new HashSet<>(), new HashSet<>(), floodingRules);
+    }
+
+    public void addFloodingProtectionRules(Set<String> filterFloodingRules, Set<String> natFloodingRules,
+            Set<String> mangleFloodingRules) {
+        try {
+            this.firewall.setAdditionalRules(filterFloodingRules, natFloodingRules, mangleFloodingRules);
+        } catch (KuraException e) {
+            logger.error("Failed to set Firewall Flooding Protection Configuration", e);
+        }
+    }
 }
